+++ examples/counter.rs
@@ -2,19 +2,13 @@ use core::time::Duration;
 
 use async_main::{async_main, LocalSpawner};
 use async_std::task::sleep;
-use pasts::{notify, prelude::*};
-
-/// List of possible events
-enum Event {
-    One(()),
-    Two(()),
-}
+use pasts::{notify, prelude::*, Loop};
 
 /// Shared state between tasks on the thread.
 struct App<'a> {
     counter: usize,
-    one: &'a mut (dyn Notify<Event = Event> + Unpin),
-    two: &'a mut (dyn Notify<Event = Event> + Unpin),
+    one: &'a mut (dyn Notify<Event = ()> + Unpin),
+    two: &'a mut (dyn Notify<Event = ()> + Unpin),
 }
 
 impl App<'_> {
@@ -36,11 +30,11 @@ impl App<'_> {
         Pending
     }
 
-    async fn run(&mut self) -> Poll {
-        match notify::select([&mut self.one, &mut self.two]).next().await {
-            Event::One(e) => self.one(e),
-            Event::Two(e) => self.two(e),
-        }
+    async fn run(&mut self) {
+        Loop::new(self)
+            .on(|s| &mut s.one, Self::one)
+            .on(|s| &mut s.two, Self::two)
+            .await;
     }
 }
 
@@ -49,9 +43,9 @@ async fn main(_spawner: LocalSpawner) {
     let sleep = |seconds| sleep(Duration::from_secs_f64(seconds));
     let mut app = App {
         counter: 0,
-        one: &mut notify::future_fn(|| Box::pin(sleep(1.0))).map(Event::One),
-        two: &mut notify::future_fn(|| Box::pin(sleep(2.0))).map(Event::Two),
+        one: &mut notify::future_fn(|| Box::pin(sleep(1.0))),
+        two: &mut notify::future_fn(|| Box::pin(sleep(2.0))),
     };
 
-    while app.run().await.is_pending() {}
+    app.run().await;
 }