+++ examples/spawn.rs
@@ -2,7 +2,7 @@ extern crate async_std;
 
 use core::time::Duration;
 
-use pasts::{prelude::*, Executor};
+use pasts::Executor;
 
 async fn sleep(seconds: f64) {
     async_std::task::sleep(Duration::from_secs_f64(seconds)).await;
@@ -12,7 +12,7 @@ fn main() {
     let executor = Executor::default();
 
     // Spawn before blocking puts the task on a queue.
-    executor.spawn(async {
+    executor.spawn_boxed(async {
         sleep(3.0).await;
         println!("3 seconds");
     });
@@ -20,11 +20,11 @@ fn main() {
     // Calling `block_on()` starting executing queued tasks.
     executor.clone().block_on(async move {
         // Spawn tasks (without being queued)
-        executor.spawn(async {
+        executor.spawn_boxed(async {
             sleep(1.0).await;
             println!("1 second");
         });
-        executor.spawn(async {
+        executor.spawn_boxed(async {
             sleep(2.0).await;
             println!("2 seconds");
         });