+++ tests/spawn.rs
@@ -1,6 +1,6 @@
 extern crate whisk;
 
-use pasts::{prelude::*, Executor};
+use pasts::Executor;
 use whisk::Channel;
 
 #[test]
@@ -10,7 +10,7 @@ fn spawn_inside_block_on() {
     let sender = channel.clone();
 
     executor.clone().block_on(async move {
-        executor.spawn(async move {
+        executor.spawn_boxed(async move {
             sender.send(0xDEADBEEFu32).await;
         });
     });